# rest-client/openapi-definition.json
{
    "components": {
        "schemas": {
...
            "CompilationL3RelationDto": {
                "properties": {
                    "compilation": {
                        "description": "Identifier of the compilation that contains the products",
                        "type": "number"
                    },
                    "productL3Src": {
                        "description": "Identifier of the product",
                        "type": "number"
                    }
                },
                "required": [
                    "compilation",
                    "productL3Src"
                ],
                "type": "object"
            },
            "ErrorDto": {
                "properties": {
                    "error": {
                        "description": "Error string",
                        "type": "string"
                    },
                    "message": {
                        "description": "Message",
                        "type": "string"
                    },
                    "statusCode": {
                        "description": "HTTP status code",
                        "type": "number"
                    }
                },
                "required": [
                    "statusCode",
                    "message",
                    "error"
                ],
                "type": "object"
            },
...
            "RelationSummaryDto": {
                "properties": {
                    "product_id": {
                        "description": "Identifier of the products",
                        "type": "object"
                    },
                    "product_name": {
                        "description": "The name of the product",
                        "type": "object"
                    },
                    "relation_id": {
                        "description": "Identifier of the relation between the products",
                        "type": "object"
                    },
                    "survey_id": {
                        "description": "Identifier of the survey/compilation that contains the products",
                        "type": "object"
                    }
                },
                "required": [
                    "relation_id",
                    "survey_id",
                    "product_id",
                    "product_name"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "access-token": {
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "contact": {
            "email": "AusSeabed@ga.gov.au",
            "name": "AusSeabed",
            "url": "http://ausseabed.gov.au/"
        },
        "description": "The API description for the Ausseabed product catalogue inventory",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "AusSeabed product catalogue",
        "version": "0.1.10"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/product-relations/compilation-to-l3": {
            "get": {
                "operationId": "ProductRelationsController_findAllL3Compilation",
                "parameters": [],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/RelationSummaryDto"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            },
            "post": {
                "operationId": "ProductRelationsController_createCompilation",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompilationL3RelationDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompilationL3Relation"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            }
        },
        "/product-relations/compilation-to-l3/compilation/{compilationId}": {
            "get": {
                "operationId": "ProductRelationsController_findConditionalCompilation",
                "parameters": [
                    {
                        "in": "path",
                        "name": "compilationId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompilationL3Relation"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the compilation"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            }
        },
        "/product-relations/compilation-to-l3/{relationId}": {
            "delete": {
                "operationId": "ProductRelationsController_deleteCompilation",
                "parameters": [
                    {
                        "in": "path",
                        "name": "relationId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the compilation"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            },
            "get": {
                "operationId": "ProductRelationsController_findOneCompilation",
                "parameters": [
                    {
                        "in": "path",
                        "name": "relationId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompilationL3Relation"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the compilation"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            },
            "put": {
                "operationId": "ProductRelationsController_updateCompilation",
                "parameters": [
                    {
                        "in": "path",
                        "name": "relationId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompilationL3RelationDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the compilation"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            }
        },
        "/product-relations/surveys-to-l0": {
            "get": {
                "operationId": "ProductRelationsController_findAllL0Survey",
                "parameters": [],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/RelationSummaryDto"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            },
            "post": {
                "operationId": "ProductRelationsController_createL0Survey",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SurveyL0RelationDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SurveyL0Relation"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            }
        },
        "/product-relations/surveys-to-l0/survey/{surveyId}": {
            "get": {
                "operationId": "ProductRelationsController_findConditionalL0Survey",
                "parameters": [
                    {
                        "in": "path",
                        "name": "surveyId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SurveyL0Relation"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            }
        },
        "/product-relations/surveys-to-l0/{relationId}": {
            "delete": {
                "operationId": "ProductRelationsController_deleteL0Survey",
                "parameters": [
                    {
                        "in": "path",
                        "name": "relationId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the compilation"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            },
            "get": {
                "operationId": "ProductRelationsController_findOneL0Survey",
                "parameters": [
                    {
                        "in": "path",
                        "name": "relationId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SurveyL0Relation"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the compilation"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            },
            "put": {
                "operationId": "ProductRelationsController_updateL0Survey",
                "parameters": [
                    {
                        "in": "path",
                        "name": "relationId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SurveyL0RelationDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "product-relations"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/rest"
        }
    ],
...
}